+++ presentation/src/main/java/com/knocklock/presentation/KnockLockNavHost.kt
@@ -8,8 +8,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
-import com.knocklock.presentation.ui.setting.SettingRoute
-import com.knocklock.presentation.ui.setting.credit.CreditRoute
+import com.knocklock.presentation.setting.SettingRoute
+import com.knocklock.presentation.setting.credit.CreditRoute
 
 @Composable
 fun KnockLockNavHost(
